Split websocket test helper in its own file
